requirements.txt
python-dotenv==0.21.1
./scripts-1.0.0-py3-none-any.whl
